Dummy clip attachment and bag interaction
Now the clip can be attached to the dummy (need to redo it slightly, i got it working with the loose clip but is a bit scuffed with the ropebag one), now in order to lift the injured to the nacelle ,the player puts the clip on the dummy, places the rotor, collides the ropebag with the rotor and presses trigger, and then uses the ratchet, also i changed a bunch of stupid variable names, lmk if you cant find something or something, idk, i dont think anyone reads these anyway, so why am i even bothering? who knows, what is the point and purpose? why are we still here? just to suffer? every night i feel my arm, and my leg, you can feel it too cant you?

;-;